uint8_t instead of int

+++ Display.cpp
@@ -69,7 +69,7 @@ void Display::resetInspHold() {
 // waveforms
 // -----------------
 void Display::updateFlowWave(float flow) {
-  int val = map(flow, FLOW_RANGE_MIN, FLOW_RANGE_MAX, GRAPH_MIN, GRAPH_MAX);
+  uint8_t val = map(flow, FLOW_RANGE_MIN, FLOW_RANGE_MAX, GRAPH_MIN, GRAPH_MAX);
   dbSerialPrintln("Flow waveform value------");
   dbSerialPrintln(val);
 //   flowWave.addValue(0, millis() % 4096 / 16);
@@ -77,7 +77,7 @@ void Display::updateFlowWave(float flow) {
 }
 
 void Display::updatePressureWave(float pressure) {
-  int val = map(pressure, PRESSURE_RANGE_MIN, PRESSURE_RANGE_MAX, GRAPH_MIN, GRAPH_MAX);
+  uint8_t val = map(pressure, PRESSURE_RANGE_MIN, PRESSURE_RANGE_MAX, GRAPH_MIN, GRAPH_MAX);
   dbSerialPrintln("Pressure waveform value------");
   dbSerialPrintln(val);
 //   pressureWave.addValue(0, millis() % 4096 / 16);
